+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 #Component: fh-mbaas
 
+##1.0.0 - 2015-01-21 - IR245 - Wei Li
 
-##0.0.1 - 2014-09-10 - IR236 - dberesford
+* 8178 Added support for life cycle management.
 
-* Ticket 7934 - fh-mbaas: create skeleton fh-mbaas service
+### Notes
+
+* The "mongo" config option should specify the connection details to the MongoDB inside the MBaas instance.
+
+* The host(s) values in those config properties:
+
+  * fhamqp
+  * fhditch
+  * fhmessaging
+  * fhstats
+  * fhredis
+
+  They are used to construct the variouse FH service environment variables for the apps when they are deployed to dynofarm. Make sure the host(s) values are accessible from inside the dyno.